# api/docs/swagger.yaml
definitions:
...
  models.CreateTransaction:
    properties:
      amount:
        type: string
      ctreate_at:
        type: string
      description:
        type: string
      sale_id:
        type: string
      source_type_enum:
        type: string
      staff_id:
        type: string
      transaction_type_enum:
        type: string
    type: object
  models.CreateTransactionStorage:
    properties:
      branch_id:
        type: string
      price:
        type: string
      product_id:
        type: string
      quantity:
        type: string
      staff_id:
        type: string
      transaction_type:
        type: string
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
...
  models.Transaction:
    properties:
      amount:
        type: string
      ctreate_at:
        type: string
      description:
        type: string
      id:
        type: string
      sale_id:
        type: string
      source_type_enum:
        type: string
      staff_id:
        type: string
      transaction_type_enum:
        type: string
    type: object
  models.TransactionStorage:
    properties:
      branch_id:
        type: string
      id:
        type: string
      price:
        type: string
      product_id:
        type: string
      quantity:
        type: string
      staff_id:
        type: string
      transaction_type:
        type: string
    type: object
...
  models.UpdateTransaction:
    properties:
      amount:
        type: string
      ctreate_at:
        type: string
      description:
        type: string
      id:
        type: string
      sale_id:
        type: string
      staff_id:
        type: string
    type: object
  models.UpdateTransactionStorage:
    properties:
      branch_id:
        type: string
      id:
        type: string
      price:
        type: string
      product_id:
        type: string
      quantity:
        type: string
      staff_id:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /transaction:
    post:
      consumes:
      - application/json
      description: create a new storage
      parameters:
      - description: transaction
        in: body
        name: basket
        schema:
          $ref: '#/definitions/models.CreateTransaction'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Creates a new storage
      tags:
      - transaction
  /transaction/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: transaction
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete transaction by ID
      tags:
      - transaction
    get:
      consumes:
      - application/json
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get transaction by ID
      tags:
      - transaction
    put:
      consumes:
      - application/json
      description: update transaction
      parameters:
      - description: transaction
        in: path
        name: id
        required: true
        type: string
      - description: transaction
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update transaction
      tags:
      - storage
  /transaction_storage:
    post:
      consumes:
      - application/json
      description: create a new storage
      parameters:
      - description: transaction_storage
        in: body
        name: basket
        schema:
          $ref: '#/definitions/models.CreateTransactionStorage'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TransactionStorage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Creates a new storage
      tags:
      - transaction_storage
  /transaction_storage/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Transaction Storage ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete transaction storage by ID
      tags:
      - transaction_storage
    get:
      consumes:
      - application/json
      parameters:
      - description: Transaction Storage ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TransactionStorage'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get transaction storage by ID
      tags:
      - transaction_storage
    put:
      consumes:
      - application/json
      description: update transaction_storage
      parameters:
      - description: transaction_storage
        in: path
        name: id
        required: true
        type: string
      - description: transaction_storage
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTransactionStorage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TransactionStorage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update transaction_storage
      tags:
      - transaction_storage
  /transaction_storages:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a list of transaction storage
      tags:
      - transaction_storage
  /transactions:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a list of transactions
      tags:
      - transaction
swagger: "2.0"